Chat implemented and working :)

+++ Travelogue/app/src/main/java/sai/developement/travelogue/fragments/AddNewTripFragment.java
@@ -293,7 +293,7 @@ public class AddNewTripFragment extends Fragment {
                                 user.setEmail((String) ((HashMap) pair.getValue()).get("email"));
                                 user.setId((String) ((HashMap) pair.getValue()).get("id"));
                                 user.setName((String) ((HashMap) pair.getValue()).get("name"));
-                                user.setAvatarId((Integer)  ((HashMap) pair.getValue()).get("avatarId"));
+                                user.setAvatarId(((Long)  ((HashMap) pair.getValue()).get("avatarId")).intValue());
                                 break;
                             }
